Created a customer with a randomly generated name, credit card.  Also created a transaction data structure to contain a customer, along with a fuel type and quantity.  Also added empty Readme file

+++ Src/tank.cpp
@@ -1,3 +1,4 @@
+#include "station.h"
 #include "tank.h"
 
 Tank::Tank()
@@ -27,7 +28,7 @@ Tank::~Tank()
 
 float Tank::decrement()
 {
-	int newVal;
+	float newVal;
 	
 	mRemain->Wait();
 	this->remaining = this->remaining - 1;
@@ -40,7 +41,7 @@ float Tank::decrement()
 
 float Tank::fill()
 {
-	int newVal;
+	float newVal;
 
 	mRemain->Wait();
 	this->remaining = MAX_CAPACITY;
